Changes float to double in distance calculations

+++ src/skeleton3d/src/repository/DistanceMatrixOperations.hpp
@@ -8,10 +8,10 @@ namespace DistanceMatrixOperations
     using TimedSkeleton = repository_data_structures::TimedSkeleton;
     template<class T> using vector = std::vector<T>;
 
-    vector<vector<float>> create_distance_observation_to_track_matrix(
+    vector<vector<double>> create_distance_observation_to_track_matrix(
         const vector<TimedSkeleton *>& observations,
         const vector<TimedSkeleton>& tracks);
-    vector<vector<bool>> find_match_over_matrix(vector<vector<float>> distance_matrix); // TODO: Impl
+    vector<vector<bool>> find_match_over_matrix(vector<vector<double>> distance_matrix); // TODO: Impl
 }
 
 #endif
